+++ application/migrations/20210412233027-create-returning.js
@@ -24,11 +24,19 @@ module.exports = {
       },
       borrowing_fk: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.INTEGER,
+        references: { // User belongsTo Company 1:1
+          model: 'borrowings',
+          key: 'id'
+        }
       },
       auth_user_fk: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.INTEGER,
+        references: { // User belongsTo Company 1:1
+          model: 'users',
+          key: 'id'
+        }
       },
       createdAt: {
         allowNull: false,